Changed data type.

+++ Section7/BankAccount.java
@@ -7,7 +7,7 @@
 public class BankAccount {
     //Creating private variables.
     private static double balance;
-    private static float withdrawal;
+    private static double withdrawal;
     private static float deposit;
     //Creating a public method.
     public BankAccount(double d, double e, double f)
